pyproject.toml
[tool.poetry]
name = "etl-project"
version = "0.1.0"
description = ""
authors = ["Vanessa Gaigher <vanessa.mat.sedu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.2"
google-cloud-bigquery = "^3.21.0"
db-dtypes = "^1.2.0"
pandera = "^0.18.3"
pyyaml = "^6.0.1"
black = "^24.4.0"
frictionless = "^5.16.1"
sqlalchemy = "^2.0.29"
dotenv-python = "^0.0.1"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"




